// src/models/requestEntradaMercanciaCajasModel.model.ts
export class RequestEntradaMercancia {
    login: string;
    data:  Data;
}

export class Data {
    produccionOF:        string;
    bodega:              string;
    cantidadCajas:       number;
    fecha:               Date;
    lote:                string;
    cantidad:            number;
    total: string;
    precioEnvase:        string;
    descripcionProducto: string;
    codProceso:          string;
    proceso:             string;
    cantxCaja:           string;
    color:               string;
    tipoResina:          string;
    codCosto:            string;
    familia:             string;
    peso:                string;
    codigoProducto:      string;
    cliente:             string;
    nombreMaquina:       string;
    precio:              string;
    ipImpresora:         string;
}